requirements.txt
Django==3.1
Pillow==7.2.0
requests==2.24.0
mime==0.1.0